+++ src/main/urls.py
@@ -56,6 +56,5 @@ urlpatterns = [
         r'^apiv1/logistica/',
         include('logistica.urls'),
     ),
-    path('grappelli/', include('grappelli.urls')),  # grappelli URLS
     path('admin/', admin.site.urls),
 ]